Lab6/src/main/java/org/example/ArrayIterator.java
package org.example;

import java.util.Iterator;
import java.util.NoSuchElementException;

public class ArrayIterator<T> implements Iterator<T> {
    private final T[] array;

    private int currentIndex = 0;

    public ArrayIterator(T[] array) {
        this.array = array;
    }
    @Override
    public boolean hasNext() {
        return currentIndex < array.length;
    }

    @Override
    public T next() {
        if (!hasNext()) {
            throw new NoSuchElementException();
        }
        return array[currentIndex++];
    }
}
